Move rate endpoint from volunteer to submission view
