+++ api/views/submission.py
@@ -4,7 +4,17 @@ from datetime import timedelta
 from typing import Union
 
 from django.conf import settings
-from django.db.models.functions import Length
+from django.db.models import Count
+from django.db.models.functions import (
+    Length,
+    TruncDate,
+    TruncDay,
+    TruncHour,
+    TruncMonth,
+    TruncSecond,
+    TruncWeek,
+    TruncYear,
+)
 from django.http import JsonResponse
 from django.shortcuts import get_object_or_404
 from django.utils import timezone
@@ -26,6 +36,7 @@ from api.authentication import BlossomApiPermission
 from api.filters import TimeFilter
 from api.helpers import validate_request
 from api.models import Source, Submission, Transcription
+from api.pagination import StandardResultsSetPagination
 from api.serializers import SubmissionSerializer
 from api.views.slack_helpers import client as slack
 from api.views.volunteer import VolunteerViewSet
@@ -201,6 +212,67 @@ class SubmissionViewSet(viewsets.ModelViewSet):
             source=source_obj,
         )
         return Response(data=self.get_serializer(queryset[:100], many=True).data)
+
+    @swagger_auto_schema(
+        operation_summary=(
+            "Retrieve a count of transcriptions for a volunteer per time frame."
+        ),
+        operation_description=(
+            "A paginated endpoint. Pass page_size to control number of results"
+            " returned, page to select a different block."
+        ),
+        manual_parameters=[
+            Parameter(
+                "time_frame",
+                "query",
+                type="string",
+                enum=["none", "hour", "day", "week", "month", "year"],
+                description="The time interval to calculate the rate by. "
+                'Must be one of "none", "hour", "day", "week", "month" or "year".'
+                'For example, "none" will return the date of every transcription '
+                'separately, while "day" will return the daily transcribing rate.',
+            ),
+            Parameter("page_size", "query", type="number"),
+            Parameter("page", "query", type="number"),
+        ],
+    )
+    @action(detail=False, methods=["get"])
+    def rate(self, request: Request) -> Response:
+        """Get the number of transcriptions the volunteer made per time frame.
+
+        IMPORTANT: To reduce the number of entries, this does not
+        include days on which the user did not make any transcriptions!
+        """
+        time_frame = request.GET.get("time_frame", "day")
+
+        trunc_dict = {
+            # Don't group the transcriptions at all
+            # TODO: Make this a true noop for transcriptions posted in the same second
+            "none": TruncSecond,
+            "hour": TruncHour,
+            "day": TruncDay,
+            # Unfortunately weeks starts on Sunday for this.
+            # There doesn't seem to be an ISO week equivalent :(
+            "week": TruncWeek,
+            "month": TruncMonth,
+            "year": TruncYear,
+        }
+
+        trunc_fn = trunc_dict.get(time_frame, TruncDate)
+
+        # https://stackoverflow.com/questions/8746014/django-group-by-date-day-month-year
+        rate = (
+            self.queryset.filter(complete_time__isnull=False)
+            .annotate(date=trunc_fn("complete_time"))
+            .values("date")
+            .annotate(count=Count("id"))
+            .values("date", "count")
+            .order_by("date")
+        )
+
+        pagination = StandardResultsSetPagination()
+        page = pagination.paginate_queryset(rate, request)
+        return pagination.get_paginated_response(page)
 
     @csrf_exempt
     @swagger_auto_schema(
